Initial implementation of local HFQL executor (#5462)
